Add files via upload
Consequently,lowering cancellation rates is both hotels ,Primary goal in order to increase their efficiency in generating revenue and for us to offer through business advice to address this problem. The analysis of hotel booking cancellations as well as other factors that have no booking on their business and yearly revenue generation are the main topics of this project.



Assumptions:

1. No unusual occurrences between 2015 and 2017 will have a substantial impact on the data used.

2. The information is still current and can be used to analyze a hotels possible plans in an efficient manner.

3. There are no unanticipated negatives to the hostel employing and advised  technique.

4. The hotels are not currently using any of the suggested solutions.

5. The biggest factor affecting the effectiveness of earning income is booking cancellations.

6. Cancellation result in vacant rooms for the booked length time.

Research Question: 1.What are the variables that affect hotel reservation  cancellations?

Hypothesis: 1. More cancellations occur when prices are higher.

2. When there is a longer waiting list, customers tend to cancel more frequently.

3. The majority of clients are coming from offline travel agents to make their reservations.